# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/balance": {
      "get": {
        "operationId": "BalanceController_getBalances",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User balances retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "fiat": {
                      "NGN": 5678910,
                      "USD": 12345.67,
                      "EUR": 8234.5,
                      "GBP": 9876.54
                    },
                    "crypto": {
                      "USDC": {
                        "amount": 1000,
                        "usdValue": 1000
                      },
                      "USDT": {
                        "amount": 500,
                        "usdValue": 500
                      },
                      "ESPEES": {
                        "amount": 2000,
                        "usdValue": 2000
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Error fetching balance"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user balances",
        "tags": [
          "balance"
        ]
      }
    },
    "/balance/send": {
      "post": {
        "operationId": "BalanceController_sendMoney",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMoneyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transfer completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Transfer completed successfully",
                    "transactionHash": "0x123..."
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid recipient details"
          },
          "401": {
            "description": "Authentication failed"
          },
          "404": {
            "description": "Recipient not found"
          },
          "422": {
            "description": "Insufficient balance for this transaction"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send money or crypto",
        "tags": [
          "balance"
        ]
      }
    },
    "/balance/request": {
      "post": {
        "operationId": "BalanceController_requestPayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment request created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Payment request created successfully",
                    "requestId": "abc123",
                    "paymentLink": "https://example.com/pay/abc123"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to create payment request"
          },
          "401": {
            "description": "Authentication failed"
          },
          "500": {
            "description": "An internal server error occurred"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request payment",
        "tags": [
          "balance"
        ]
      }
    },
    "/balance/withdraw": {
      "post": {
        "operationId": "BalanceController_withdraw",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Withdrawal request submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Withdrawal request submitted successfully. You will receive an email confirmation shortly.",
                    "withdrawalId": "wd_123456",
                    "amount": 500,
                    "currency": "USD",
                    "status": "PENDING"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Withdraw funds",
        "tags": [
          "balance"
        ]
      }
    },
    "/balance/convert": {
      "post": {
        "operationId": "BalanceController_convertCurrency",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConvertCurrencyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Currency conversion successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "convertedAmount": {
                      "type": "number",
                      "example": 1800
                    },
                    "rate": {
                      "type": "number",
                      "example": 1800
                    },
                    "from": {
                      "type": "string",
                      "example": "ESP"
                    },
                    "to": {
                      "type": "string",
                      "example": "NGN"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Convert between currencies",
        "tags": [
          "balance"
        ]
      },
      "get": {
        "operationId": "BalanceController_convertAmount",
        "parameters": [
          {
            "name": "amount",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "balance"
        ]
      }
    },
    "/balance/rates": {
      "get": {
        "operationId": "BalanceController_getExchangeRates",
        "parameters": [
          {
            "name": "base",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exchange rates retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "base": "USD",
                    "rates": {
                      "EUR": 0.85,
                      "GBP": 0.75,
                      "NGN": 411.5,
                      "USDC": 1,
                      "USDT": 1,
                      "ESPEES": 1
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current exchange rates",
        "tags": [
          "balance"
        ]
      }
    },
    "/balance/crypto/price": {
      "get": {
        "operationId": "BalanceController_getCryptoPrice",
        "parameters": [
          {
            "name": "symbol",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "balance"
        ]
      }
    },
    "/balance/activity": {
      "get": {
        "operationId": "BalanceController_getRecentActivity",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Recent activity retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "type": "RECEIVED",
                      "currency": "USDC",
                      "amount": 100,
                      "timestamp": "2024-03-20T10:00:00Z",
                      "description": "Payment received"
                    },
                    {
                      "type": "SENT",
                      "currency": "USD",
                      "amount": 50,
                      "timestamp": "2024-03-20T07:00:00Z",
                      "description": "Transfer to john@example.com"
                    },
                    {
                      "type": "WITHDRAWAL",
                      "currency": "EUR",
                      "amount": 200,
                      "timestamp": "2024-03-19T15:30:00Z",
                      "description": "Bank withdrawal"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed"
          },
          "500": {
            "description": "An internal server error occurred"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get recent balance activity",
        "tags": [
          "balance"
        ]
      }
    },
...
  },
  "info": {
    "title": "EspeePay API",
    "description": "The EspeePay API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SendMoneyDto": {
        "type": "object",
        "properties": {
          "assetType": {
            "type": "string",
            "enum": [
              "FIAT",
              "CRYPTO"
            ],
            "example": "FIAT",
            "description": "Type of asset to send"
          },
          "recipientAddress": {
            "type": "string",
            "example": "user@example.com",
            "description": "Recipient email or wallet address"
          },
          "amount": {
            "type": "number",
            "example": 1000,
            "description": "Amount to send"
          },
          "currency": {
            "type": "string",
            "enum": [
              "USD",
              "GBP",
              "EUR",
              "NGN",
              "USDC",
              "USDT",
              "ESPEES"
            ],
            "example": "USD",
            "description": "Currency for the transfer"
          },
          "note": {
            "type": "string",
            "example": "Payment for services",
            "description": "Optional note for the transaction"
          }
        },
        "required": [
          "assetType",
          "recipientAddress",
          "amount",
          "currency"
        ]
      },
      "RequestPaymentDto": {
        "type": "object",
        "properties": {
          "currency": {
            "type": "string",
            "enum": [
              "USD",
              "GBP",
              "EUR",
              "NGN",
              "USDC",
              "USDT",
              "ESPEES"
            ],
            "example": "USD",
            "description": "Currency for payment request"
          },
          "amount": {
            "type": "number",
            "example": 1000,
            "description": "Amount to request"
          },
          "payerEmail": {
            "type": "string",
            "example": "payer@example.com",
            "description": "Email of the person who should pay"
          },
          "description": {
            "type": "string",
            "example": "Invoice for February services",
            "description": "Description of the payment request"
          }
        },
        "required": [
          "currency",
          "amount",
          "payerEmail",
          "description"
        ]
      },
      "WithdrawDto": {
        "type": "object",
        "properties": {
          "currency": {
            "type": "string",
            "enum": [
              "USD",
              "GBP",
              "EUR",
              "NGN",
              "USDC",
              "USDT",
              "ESPEES"
            ],
            "example": "USD",
            "description": "Currency to withdraw"
          },
          "amount": {
            "type": "number",
            "example": 500,
            "description": "Amount to withdraw"
          },
          "accountNameOrAddress": {
            "type": "string",
            "example": "John Doe",
            "description": "Bank account holder name (for fiat) or crypto address (for crypto)"
          },
          "accountNumber": {
            "type": "string",
            "example": "1234567890",
            "description": "Bank account number (required for fiat withdrawals)"
          },
          "bankName": {
            "type": "string",
            "example": "First Bank of Nigeria",
            "description": "Bank name (required for fiat withdrawals)"
          },
          "bankCode": {
            "type": "string",
            "example": "12345",
            "description": "Bank sort code or routing number (required for some fiat withdrawals)"
          }
        },
        "required": [
          "currency",
          "amount",
          "accountNameOrAddress"
        ]
      },
      "ConvertCurrencyDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "description": "Amount to convert",
            "example": 100
          },
          "from": {
            "type": "string",
            "enum": [
              "USD",
              "GBP",
              "EUR",
              "NGN",
              "USDC",
              "USDT",
              "ESPEES"
            ],
            "description": "Source currency",
            "example": "USD"
          },
          "to": {
            "type": "string",
            "enum": [
              "USD",
              "GBP",
              "EUR",
              "NGN",
              "USDC",
              "USDT",
              "ESPEES"
            ],
            "description": "Target currency",
            "example": "EUR"
          }
        },
        "required": [
          "amount",
          "from",
          "to"
        ]
      },
...
    }
  }
}